feat: 支持GitHub Re-run #3676

+++ src/backend/ci/core/plugin/api-plugin/src/main/kotlin/com/tencent/devops/plugin/api/pojo/GithubCheckRun.kt
@@ -34,6 +34,6 @@ data class GithubCheckRun(
     val buildNumber: Int,
     val repositoryConfig: RepositoryConfig,
     val commitId: String,
-    val checkRunId: Int,
+    val checkRunId: Long,
     val checkRunName: String? = null
 )
